OpenCollar - rlvsit.lsl:
The SitNow command scans for active objects and excludes those named "Object". When one is skipped, though, it is not added to the list of objects used later for sitting, however the index used in the dialog is the loop index which also counts skipped numbers. This results in desynchronization of the dialog with the list indices, resulting in selecting the wrong object to sit on.

For example, say the sensor detects 4 objects, called A, Object, C and D. Then the list will contain A, C, D in positions 0, 1, and 2 respectively, and the dialog will contain buttons ( 1 ), ( 3 ) and ( 4 ) claiming they are for A, C, and D respectively. Pressing ( 1 ) correctly selects index 0 which is A. However, pressing ( 3 ) invokes index 2 which is D instead of C, and pressing ( 4 ) invokes index 3 which does not exist at all.

This mess is fixed by keeping a separately running index which is incremented only when objects are actually added to the list.

And since I'm here, global substitution of SiTPrompt with SitPrompt.