rlvrelay.lsl: Fix 2 bugs and optimize
Revert the order of Min and non-Min when restoring settings. Fixes issue 1342 and issue 1224.

Add key sanitization function. Fixes issue 1388.

Due to the code added in the fix of 1388 and the scarce memory available for normal relay operation, an agressive overall memory optimization is also applied to regain the lost memory.