Multiple settings per link message
Added the capability to save, request and delete multiple settings per single link message.

To store settings, use a HTTPDB_SAVE message with name/value pairs separated by a pipe character ("|", SEPARATOR in the script), e.g. "name1=value1|name2=value2". To retrieve or delete settings, use the corresponding message with names separated by "|".

The response to a HTTPDB_REQUEST message may be either a HTTPDB_RESPONSE or a HTTBDB_EMPTY message, or both, depending on the stored values. They are formatted similar to the SAVE and REQUEST messages.

The "Test - settings.lsl" script is a first try to encourage automatic testing.