First commit to the "auth once" branch. This branch involves:  - removing SUBMENU LMs, now replaced by authed LM's COMMAND_XXX 	with prefix "menu "  - adding another parameter to dialog helper LMs to carry auth data As a result, COMMAND_NOAUTH is now only used by auth plugin and input plugins such as menu (touch event) and listener. Also, SUBMENU is now deprecated.
Another part of the work was to make menu calls synchronous
everywhere it was now posssible, since auth is preserved
(so way less "remenu" stuff).
COMMAND_XXX handlers are usually moved to a function UserCommand()
which can be called from anywhere in the script to call internal
commands instead of having to send yet another LM to all scripts.