First implementation of the new touch input system.
There are 2 subsystems:
- permanent actions which are collar chat command that can be added in the prim descriptions. "menu" is the default if nothing is specified. It is
also possible to disable the default action.
- temporary actions which will be enabled only for the next click, and can be used like dialogs, except the user "presses" a prim instead of a
button. Temporary actions override all other actions, but only for the next click.
Both subsystems can address touch_start and touch_end events separately.

As an example, the color plugin has been modified to make it possible for the user to pick the part to recolor using this second subsystem.