multiple settings per link message
lsl/OpenCollar - settings.lsl: enable request for multiple settings with a single link message. This enables the auth script to retrieve settings with a single message instead of eight. This is implemented in the settings script only (enabling the feature), others are to follow.

test/Test - settings.lsl: regression test for the settings script