Medea Destiny in-world; lock
Medea:

New feature: Added support for setting in defaultsettings notecard
User_lock=locksound~<value> and User_lock=unlocksound~<value> accepting
either a UUID or a sound in inventory or "default", which restores the
sounds to the default setting: User_lock=locksound~default . Shall we
put something in Options menu to set this? Or keep it as a designer's
thing? Or play a sound called "lock" if found in inventory regardless of
settings?

Other enchancements: Memory limit set to llGetUsedMemory (after building
list)+6000 should enough memory, but THIS NEEDS WATCHING! Minor
efficiency improvement to SetLockElementAlpha (using g_iLocked and
!g_iLocked for alpha values instead of fAlpha). Removed unnecessary
on_rez event. Fixed GetPSTDate so that conversion on the 1st of the
month before 8am UST doesn't return the 0th of the month. Fix foolishly
does not account for centuries not being leap years, yes, yes, I know I
was lazy, but we've got 86 years to fix this.