Track and Clear listener handles of ":ping"ed avatars, rather than switching states.
Rather than switching states to clear the listeners after a ping for subs, (and therefore having to reinstate the ones we need), this change tracks the listener handles and removes them after 5 seconds.

I also believe this is cleaner for future maintainability of default/state_entry(), which many would assume only occurs on script reset.