Multisearch for RLV Folders
This feature is by Medea Destiny:

improved chat functionality for attaching folders. Now accepts two-level
search terms with a /. Example: <prefix>&&gag/rr will find the first
folder with "gag" in the title just as <prefix>&&gag would now, but
instead of attaching that, will look for a subfolder below that with
"rr" in the title, and attach that folder instead.

additional improvement: search with no result caused "no folder found"
message to be sent to kID, which is llGetOwner() in listen event,
changed to g_kAsyncMenuUser to return error message to person issuing
the search command.