make wearer set leash store auth as COMMAND_GROUP
A situation was discovered where a wearer who was their own owner (or secowner) with public access set could leash themselves to an unwillling person, and that person be unable to unleash themselves.  This change forces any leash set by the wearer to be stored with a group level auth, which can be broken in public access mode.