disable debugging and profiling
Mouse added a very clever way to disable/enable debugging. It's
efficient, quick and a chimpanzee would be able to use this method.
However, you should keep in mind that when you let the compiler bop you
from line, it's a genuinely bad idea to do this in a script nested
inside of an object you are currently wearing. The asset will go
hay-wire eventually and you might lose n hours of work unless you
regularly save back to your local drive. This has nothing to do with the
debugging, it's just if saving a nested script in quicker succession
than the database can keep track of it.

I've created a short ritual for myself, you do it your way but maybe
this might help you too:

1. I decide that the script is now production ready
2. I load the contents of my collar
3. I drag the script I had been working on to my inventory
4. I close all scripts to prevent mixing stuff up
5. I open the script in my inventory
6. I follow the pattern to disable debugging
7. I save to my local drive when the compiler is happy
8. I open my object once again and load the production ready script from
my local drive

This is happening in motion with guideline #331 and committed for future
reference.