create rlvview spare part (read description!)
This is a copy of rlvstuff with only the Camera/View specific features,
however, there are also leftovers from sit commands etc.. so it will
have to be cleaned up before putting it into a collar.

For now, we can work on this source and improve the view section to
merge it back into rlvstuff at a later point or we can work on our
existing optional Camera app.

Camera/View features as present in RLV 2.9 have to remain optional until
the majority of viewers provide this or a similar feature set and there
is no rush for this.