standardize REQUIRED bundle
This installer bundle is called "Core" and tagged REQUIRED. When we
perform an installation, bundles tagged REQUIRED will always be
installed and cannot be unticked in selection menus. A good choice for
this type of bundle could be the basic elements to make the device
operational.

Example:
ITEM|license
ITEM|readme
SCRIPT|sys
SCRIPT|com
SCRIPT|db