standardize INSTALL bundle
This installer bundle is called "Utility" and tagged INSTALL. When we
perform an installation, bundles tagged INSTALL will be pre-selected to
be installed as "standard" but can be unticked in selection menus if so
desired by the user. A good choice for this type of bundle could be the
kind of features the user usually almost always expects to have in the
device.

Example:
SCRIPT|monitor
SCRIPT|debug
SCRIPT|archive