standardize REMOVE bundle
This installer bundle is called "App" and tagged REMOVE. When we perform
an installation, bundles tagged REMOVE must be selected first so to be
installed. A good choice for this type of bundle could be all sort of
optional applications depending on the users individual need and taste.

Example:
SCRIPT|chat
SCRIPT|game
SCRIPT|office