add list of all known authors and correct dates
While this may not be necessary with a copyleft license, I think it had
to be done if we imagine that on SL and OpenSim a script can only ever
have a single "creator" and considering the nature of the platform,
somebody curious who looks inside can at least see that these scripts
are composed by many people together, which is what I'm trying to
convey.

I also received an e-mail already a while ago that Google Code will shut
down and our legacy SVN trunks are read-only by now and probably soon
won't be accessibly at all.

Maybe we could merge Nandana's initial ocupdater repo as a sub-directory
into our new communal repository and branch off at a 3.701 point to
preserve this important phase in the history of our little, big project
as well.