add info feature for release notes etc.
A card ".info" is added to the installer contents. The first line of
this card is an url to be used with the loadurl dialog and the following
lines are a short blurb we print after installations are completed. If
no ".info" card is present, we skip the link and info blurb.