change ~version name and text to a full name
Instead of adding the version number from the "~version" card to the
object's name we change the objects name to the full text given in a now
called ".name" card. This card will set the first line as object name
and for hover text, will do a linebreak after a space-hyphen-space " -
".

thanks @GarvinTwine