drop the bass
This will play a bass drum drop when owners or trusted are added.

The sample is from a Roland TR-808 Rhythm Composer with Bass Drum set to
minimum tuning and maximum decay.

Attribution goes to Michael Fischer who generated this legendary drop
and other samples from a real TR-808 in September 1994 on a Macintosh
Quadra 660AV with SoundEdit 16.