don't position the hud automatically
It would be useful on first wear, however, this will in many cases pop
the HUD of the visible screen once first minimized or maximized. I will
not remove these bits in case somebody wants to do it properly as it is
done in the new Remote and only comment them. Keep in mind the behaviour
for visible screen differs in between viewers, resolutions and operating
systems.