build opencollar six
This is the OpenCollar Six Release Candidate.

I added g_sDevStage to indicate the current development stage of the
product. Please use whatever works, legacy OpenCollar had DEV/BETA/RC, I
also think MVP is a good to use. When we are golden this can be an empty
string.

The Remote and AO only indicate their development stage in the fancy
version and object name as these are all the user really sees.

The OpenCollar Remote is also a Release Candidate and the OpenCollar AO
is the minimum viable product, fully compatible and functional, however
should be all means become a server side AO when golden.