don't turn RLV off on failed handshake
There would be better ways to solve this if we hadn't ran out of time.
What happens here is that after a failed flurry of handshakes, we retain
RLV=on~1 and inform the user with a notify. Also the menu has new
informative text and an ON as well as an OFF button. There is also a
short notify to tell when RLV is starting if using the command or
pushing the button (because we can't easily re-menu in this case for
obvious reasons).

A proper solution would be to only give a RLV menu after successful
handshake, whether we have RLV active in the viewer or not doesn't
matter. So the problem remains that users come online with RLV disabled
in their viewers but have a menu full of features that cannot work
available for approx. 2 minutes. What is solved though is the case where
the user logs on with two or more different viewers and then forgets to
turn RLV on once again and erroneously believes she has force teleport
failing.

Unless I made a mistake somewhere, this improves the situation slightly
but doesn't solve the problem properly.