update checkboxes, menu prompts and empty buttons
I swapped most of the "ballot box with x" for a "ballot box with check",
in many cases the "ballot box with x" is used now as "radio buttons" but
it can also be just where it is visually more appealing to alternate the
two types.

Then I also changed the "empty buttons" to a button with a hyphen to
resemble more of an empty socket rather than a blank button which seems
to have made a less confusing impression on those users who gave active
feedback on that minor detail.

The Settings menu prompt is clear now, I think it is best if the man
page for this part of the device is read rather than half-baked info in
the menu prompt. Menu prompts aren't manual pages. :speak_no_evil: