separate tokens with tilde character
Theme tokens have to be separated with the tilde "~" character, similar
to how we do it in settings in general.

NOTE: If a token in the middle of a theme is empty, we use two periods
",," like:

Bolt~5748decc-f629-461c-9a36-a35a221fe21f~,,~low

In this example the Bolt element in the theme has a texture, doesn't
touch the color, has low shine and ignores the glow.