compress header; make space
Right. We will need a bit more space there to be able to put a better
step-by-step guide on adding this script. I think the script name is
also not ideal to be dropped into an object with large amounts of
content.

I'm going to rename it with a dot in front so it will bop to top in
object contents.