don't fetch default bookmarks
Once upon a time we would ship our collar with a current landmark, but
when we had less (land tier) funding, yet still a whole lot of traffic,
and had to move our home around more frequently we needed a solution
that could refresh this important landmark in between update releases.

This was done with simple HTTP requests to fetch recent SLURLs.

Meanwhile we don't have to rely on that anymore because we have a stable
website (www.opencollar.at) that provides rich information we can update
without having to write to the collar, and our bookmarks app can ship
pure and without any preset or default bookmarks/landmarks saved.

Cutting down on HTTP requests is a good thing. :)