scan for lock elements, lay out basic functionality
First we need to scan for elements that resemble the lock, that would be
"ClosedLock" and "OpenLock" and for using the script in very old collars
also just "Lock".

We also need to watch for changes in the link-set (before I forget about
it again), so that goes into the link message event as a first.

Then we have to take care of the actual command handling and do that in
the link message event (we really don't need a commands function just
for that). At this point let's make sure that the usual expectations are
fulfilled, that means informing when the use is not authorized (no
access), playing a noise when the lock is triggered and prepare a
function for revealing/hiding the lock depending on state (which we want
to learn about in the setting response event).

Very important: We have to reset the script when a device has a new
owner (wearer).