add function to show/hide the lock, handle glow
Things are getting a bit cumbersome when we have to hide and reveal our
lock elements. We need a special function just to handle glow because a
prim (100% set transparency rather than a 100% transparent texture) will
continue to glow if made merely "invisible". The glow values have to be
fetched, stored, and be restored once the lock shows up.

We also have to check whether the device is hidden on rez and color
change, and store it in the "hidden" variable. (That prevents the lock
to show up when the device is locked while stealthed.)

The reboot command we want to add as well of course.

PS: RLV is relevant here. I'm not convinced to fully rely on oc_rlvsys
(as seen earlier when I used a plain llOwnerSay also to set the detach
restriction) for the most important lock, sending a LM comes with a
delay, if so a tiny one, but if anyone feels that "double" the
restrictions are too much, just remove them.