Particles in titler
Needs a new prim with appropriate desc (and should degrade gracefully in its absence).  Not hooked into menu system yet; has to be done with # title particle <KEY>.  And that's another thing it doesn't have yet: shorthand names for some standard particles (I've made and found a few; someone with better graphic skills could do good things).

This is cooler than it sounds: the particles are not flying around; they're stationary graphics hanging in the air.  So they could be words like hovertext, but with cool graphics and coloring and design and everything.  One notable issue: since the particles have to be *at* a prim, there has to be a prim overhead for them, which means a prim linked to but fairly far from the collar, so it moves pretty widely with small movements of the head.

See what you think,