Fix oc_leash: Height fix and settings fix
The former patch to oc_settings causes a infinite loop. This fix will now prevent a infinite loop of dispatching the leash settings to LM_SETTING_SAVE. Only issues this linked message IF it detects that the user has changed the settings or on object_rez, it won't do it in the link_message LM_SETTING_RESPONSE any longer.