Update oc_relay.lsl
Changes to auth function:
1. In landowner trusted mode, we revert to ask mode if the landowner is the collar wearer.
2. When the collar is in ask mode and the object is owned by someone on the collar owner list, collar owner trumps ask mode and auth is auto. However if the object issuing the command is owned by someone on the owner list, that can create issues. 
a. We do not trump ask mode if the owner of the object is the wearer of the collar. You shouldn't be unable to be in ask mode for your own devices!
b. If the actual avatar owning the object is not in the sim we shouldn't presume their auth, so we only trump ask mode if they are present.