Updated limits for owners, oc_settings
Updated the limits in the oc_settings script in order to get away from the extremely limiting limits that were set 4 years ago.
The edits makes the adding of owners from notecard "unsafe" in that you may add too many and crash.
But as it was before, only 3 primary owners could be added from settings notecard, only 15 trusted and only 9 blocked.
This is despite the auth script allowing 28 of any combination of these added.
If you add 10 owners to your collar and then press to save your settings and put them into the settings notecard, you'll only receive 3 of them and the save won't have worked as you expected.
Removing the limits makes the saving and loading work as you expect.
There is indeed a risk that you may send too many users to the auth script and potentially crash it in case the auth script doesn't handle that. But this is only possible if you manually add users to the notecard yourself, without using the Print feature.
Thus that can be considered an advanced feature where the user will have to blame themselves for messing with the collar without knowing what they are doing in favor of making it work as expected for the normal users.
Furthermore, auth should be the one handling its own memory, settings shouldn't handle auth's memory for it.