Add ability to grab the final dialog menu text or the button list from a custom link message
The dialog performs custom tasks on a button list or on the menu text. This relays the information on a custom channel which is not being listened to by other scripts. This allows plugins to use this information.
TODO: Maybe add the actual page number to this output at a later time