Fix bug in settings: auth no longer limited to X number of owner,trust,block
Additionally; It would appear that regardless of this bug, or intended feature, notecard lines are capped at 255 bytes. To reflect this limitation, a advanced configuration option is added with this commit. append mode. If the token does not exist, this mode operates the same as set.

An example of this mode is: 
TOKEN=var~value1
TOKEN+var~value2

This results in the final value: token=var~value1,value2

Additionally, we realize not everyone may want a comma added automatically with this mode, so just set this at the top of your .settings to prevent commas with append mode:

SETTINGS=nocomma~1