Installer fix - update to oc_update_shim
The boot speedup fixes in PR #667 were causing problems for installation with collars with a heavy settings load. In investigating I came up with a few issues with the current update process that this fix at least patches around.
1. The timer was set to 1 second, but timer checks in the timer event assumed a 2 second timer, which caused timeouts at 5, 15, and 25 seconds to be duplicated at 6,16 and 26 seconds.
2. The timer was requesting settings be sent again if the settings=sent message hadn't been received yet. With settings being sent at 1 per quarter second (max - assuming no lag!) and the first timer trigger after just 5 seconds, this was getting triggered and the settings script was being hammered with requests to resend settings. I have changed it so the timer only triggers when NO setting has been received during the wait, which makes receiving settings faster and more reliable.
3. The signal to reboot the collar after the installation had happened was being sent before the update shim restores settings to the oc_settings script. This meant that those settings being properly handled was timing-dependant, and the faster boot process didn't like it. I have moved the reboot command to AFTER the settings get sent to the new oc_settings script, to ensure the collar reboots with the new settings.

All this has seemed reliable in testing, while the previous version was not saving settings with the 8.2 faster boot times, on collars with a lot of settings. Overall the install time tends to be slightly longer, perhaps 10 seconds or so, but that's a lot better than having the collar boot 30 seconds more slowly each time it is attached, which seems to be the alternative.