only show the Weld menu button when first bit is set
By checking for the first bit rather then the whole integer, the same functionality remains as before for a TRUE value.  
But if the first bit isn't set but others are, then the Weld option is removed from the main menu.
This can then simply be configured in .settings as 'intern=weld~2' for hiding the button.

so for 1 or 0  it works the same.
for 2 the Button hides.