Speedup improvements
The ALIVE signal is going to be needed to speedup the bootup sequence in oc_states.
No longer waiting for initialize, if we got an ALL request we already know the receiving scripts are ready.
Using the timer as a delay makes it only delay after a reboot/reset. Making it faster then a hardcoded delay on initialize while still respecting the time needed when needed.