Update oc_core.lsl
Fixes for issue #774. This is not a complete solution, but checks agent key match g_kWearer rather than checking auth level is CMD_WEARER throughout the script. This  should fix the issues discussed in the #774, but there are other points in a few other scripts that will have to be addressed before closing that issue.