Installer update quick maybe works who knows thing
Look, I've just done this in a few minutes without properly poking around because Roan was DESPERATE. I think this is gonna change the "Skip" button to "Keep" in cases where it's asking to remove or retain an optional script that's already in the collar, but I haven't tested this at all and I only had a few minutes to poke at the installer so there has been no proper logic testing or ANYTHING.

So basically I think it's gonna work, but it might not. I dunno! It's here because Roan wanted it fast. I've done it fast. It might work. Try it out and see! Yay fun. :P