fix some wrongs in DIALOG, DIALOG_RESPONSE & DIALOG_TIMEOUT
added check "iMenuIndex" in  DIALOG_TOMEOUT  (  ~iMenuIndex equals
iMenuIndex != -1  ) for proper removal stride from list g_lMenuIDs

there are many mistakes!
in some scripts it was forgotten to specify a constant: g_iMenuStride=3
important! this constant determines the size of the strides in the list
g_lMenuIDs

since the size of the strides in the list is always 3, there is a reason
to replace it with a numerical constant in relevant functions

in Dialog() it looks like this:
if (~iIndex) g_lMenuIDs = llListReplaceList(g_lMenuIDs, [kID, kMenuID,
sName], iIndex, iIndex + g_iMenuStride - 1);
will be equal to:
if (~iIndex) g_lMenuIDs = llListReplaceList(g_lMenuIDs, [kID, kMenuID,
sName], iIndex, iIndex + 2);

in "DIALOG_SESPONSE" and "DIALOG_TIMEOUT" it looks like this:
g_lMenuIDs = llDeleteSubList(g_lMenuIDs, iMenuIndex-1, iMenuIndex+1);